refactor: realocate components to ui section and separate them
